Fix problem with keyboard staying open on Android after screen change (#603)
This is a follow up fix after #592 where we attempted to fix the issue while introducing a regression in the way autoFocus is handled. That is, in case there was a text field with an auto focus in the next screen, the logic we added for screen dismissal would cause it to immediately lose focus.

This fix moves the logic responsible for keyboard force-dismiss to execute earlier – at the moment when we attempt to detach the previous screen. This way we can still rely on the focused child property to point to a correct children. Note that before, detaching a view would result in the child losing focus but the keyboard would stay open. We now use this fact and can verify that the focused children is the one being removed, and only in that case we ask system to hide the keyboard.