feat: add RTL support (#580)
Added RTL support for both platforms. In order to use rtl on Android, you have to add android:supportsRtl="true" in the manifest of your app.
On Android, if you set the above flag in the manifest, the orientation changes without the need to do it programmatically. On iOS, the direction defaults to ltr, and only way to change the direction is to do it programmatically.