Fix relayout loop on header show/hide change (#712)
## Description

This PR fixes a rarely occuring problem with a re-layout loop on header visibility change. The problem would surface in situation when we have a bottom tabs container inside of a stack, then we'd update header visibility such that it is hidden on some tabs and not on the other. As a result in some cases the screens code would end up re-layouting the screen with header disabled. The loop was caused by the code in RNSScreen where we call `updateBounds` as a result of native layout but also notify uimanager about view bounds change with `setSize` method. 

This PR breaks the loop by preventing screen container (not in stack) from calling `updateBounds`. We only need to notify uimanager about container size when the screen is rendered under navigation controller or as a native modal.

## Changes

We update the logic in RNScreen to prevent `uiManager setSize` call via `updateBounds` to be executed unless the screen is rendered directly by navigation controller or as a native modal. Otherwise the size should be controller by react-native layout and we should never modify the size in uimanager.

## Test code and steps to reproduce

Apparently I have not been able to find an easy reproducable case for this issue. But can confirm it has fixed an issue for one of the apps I'm working on.