Use ReactNativeStyleAttributes to process fontFamily (#823)
Expo is moving away from using the .expo extension to identify "managed" apps in SDK 41, so we are patching libraries that currently depend on the extension to fork behavior.

The solution used here is more generic - we use the same process function that is set via StyleSheet.setStyleAttributePreprocessor [(the method we use in Expo managed apps to scope fonts)](expo/expo@fb198eb/packages/expo/src/Expo.fx.tsx#L36-L39. In the (unlikely) case that other (non-Expo) folks are depending on this behavior it will also work out of the box in native-stack. I'm not crazy about doing a deep import into React Native like this, but there is no public API for getting StyleSheet preprocessors from React Native.