feat: add headers to modal (#821)
Added option for achieving headers in modal on iOS. The change is based on adding additional ScreenStack to every modal Screen component and adding the HeaderConfig component to upper or lower Screen based on the stackPresentation to ensure proper header transitions. Modal headers will not have the back button on iOS, since the header comes from the nested stack there and has no previous screen.