feat: remove adding to back stack logic on Android (#1148)
PR removing the behavior of attaching listener for backStack changes on Android. Based on stackoverflow.com/a/60560844/13006595 it looks like it is not available while using synchronous Fragment updates via commitNow. Also, it was not used in react-navigation since it provides its own listener for hardware back button on Android.

It can be considered BREAKING CHANGE if you use this library outside of react-navigation