feat: add transition duration on iOS (#1259)
## Description

Added transition duration on iOS requested in https://github.com/software-mansion/react-native-screens/discussions/1171.

The prop changes the duration of `slide_from_bottom`, `fade_from_bottom`, `fade` and `simple_push` transitions on iOS. The duration of `default` and `flip` transitions isn't customizable.

Such prop could be also added on `Android`, but each of fragments have its own transition duration, so the way it should be applied would need to be thought over. One of options is to set the duration of currently added or currently removed fragment on both of them.